src/apis/competiotion-api/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.27.0
aio-pika==9.4.0
psycopg2-binary==2.9.11
django-prometheus==2.4.1
python-logging-loki==0.3.1
minio==7.2.7
filetype==1.2.0
